Use market_game schema for betting tables
